requirements.txt
backports-abc==0.5
backports.shutil-get-terminal-size==1.0.0
bleach==3.0.2
certifi==2018.10.15
configparser==3.5.0
decorator==4.1.2
dj-database-url==0.4.2
Django==1.11.6
django-configurations==2.0
django-debug-toolbar==1.8
django-dotenv==1.4.1
django-extensions==1.9.6
django-filter==1.1.0
django-jet==1.0.6
djangorestframework==3.7.3
entrypoints==0.2.3
enum34==1.1.6
functools32==3.2.3.post2
futures==3.2.0
gunicorn==19.7.1
ipaddress==1.0.22
ipykernel==4.10.0
ipython==5.8.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
Jinja2==2.10
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==5.2.0
jupyter-core==4.4.0
Markdown==2.6.9
MarkupSafe==1.0
mistune==0.8.4
mkl-fft==1.0.6
mkl-random==1.0.1
nbconvert==5.3.1
nbformat==4.4.0
networkx==2.0
newrelic==2.96.0.80
nltk==3.2.4
notebook==5.7.0
numpy==1.15.3
pandas==0.23.4
pandocfilters==1.4.2
pathlib2==2.3.2
pexpect==4.6.0
pickleshare==0.7.5
prometheus-client==0.4.2
prompt-toolkit==1.0.15
psycopg2==2.7.3.2
ptyprocess==0.6.0
Pygments==2.2.0
python-dateutil==2.7.3
pytz==2017.3
pyzmq==17.1.2
qtconsole==4.4.2
scandir==1.9.0
scipy==1.0.0
Send2Trash==1.5.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.11.0
sqlparse==0.2.4
summa==0.0.7
terminado==0.8.1
testpath==0.4.2
tornado==5.1.1
traitlets==4.3.2
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.12.2
widgetsnbextension==3.4.2
